+++ client/src/App.jsx
@@ -18,25 +18,25 @@ import Landing from "./components/Landing";
 import Author from "./components/Author";
 import BookDetails from "./components/BookDetails";
 import AuthorDetails from "./components/AuthorDetails";
-import Add from "./components/Add";
-import AddBook from "./components/AddBook";
-import AddAuthor from "./components/AddAuthor";
+import Add from "./components/Admin/Add";
+import AddBook from "./components/Admin/AddBook";
+import AddAuthor from "./components/Admin/AddAuthor";
 import BottomBar from "./components/BottomBar";
-import { Admin } from "./components/Admin";
+import { Admin } from "./components/Admin/Admin";
 import UserPerfil from "./components/UserPerfil";
-import DeleteData from "./components/DeleteData";
-import Put from "./components/Put";
-import PutAuthor from "./components/PutAuthor";
-import PutBook from "./components/PutBook";
-import PutAuthorID from "./components/PutAuthorID";
-import PutBookID from "./components/PutBookID";
+import DeleteData from "./components/Admin/DeleteData";
+import Put from "./components/Admin/Put";
+import PutAuthor from "./components/Admin/PutAuthor";
+import PutBook from "./components/Admin/PutBook";
+import PutAuthorID from "./components/Admin/PutAuthorID";
+import putBookId from "./components/Admin/PutBookID"; 
 import ProtectedRoute from "./components/ProtectedRoute";
-import Stock from "./components/Stock";
+import Stock from "./components/Admin/Stock";
 import LogInButton from "./components/LogIn";
 import LogOutButton from "./components/LogOut";
 import { useAuth0 } from "@auth0/auth0-react";
-import DeleteAuthor from "./components/DeleteAuthor";
-import DeleteBook from "./components/DeleteBook";
+import DeleteAuthor from "./components/Admin/DeleteAuthor";
+import DeleteBook from "./components/Admin/DeleteBook";
 import AdminPro from "./components/AdminPro/AdminPro";
 import UserDatos from "./components/UserDatos";
 import UserSuscripcion from "./components/UserSuscripcion";
@@ -44,15 +44,15 @@ import ShoopingCart from "./components/ShoppingCart";
 //import AdminUsers from "./components/AdminUsers";
 import AdminOrders from "./components/AdminOrders";
 import AdminCarousel from "./components/AdminCarousel";
-import NavBarAdmin from "./components/NavBarAdmin";
-import StockTable from "./components/StockTable";
-import AdminUsers2 from "./components/AdminUsers2";
+import NavBarAdmin from "./components/Admin/NavBarAdmin";
+import StockTable from "./components/Admin/StockTable";
+import AdminUsers2 from "./components/Admin/AdminUsers2";
 import CreateAdmin from "./components/AdminPro/CreateAdmin";
 import UserFav from "./components/UserFav";
 import { AdminProProfile } from "./components/AdminPro/AdminProProfile";
-import AdminUserProfile from "./components/AdminUserProfile";
+import AdminUserProfile from "./components/Admin/AdminUserProfile";
 import ProtectedRouteBan from "./components/ProtectedRouteBan";
-import Banned from "./components/Banned";
+import Banned from "./components/Admin/Banned";
 import UserNavBar from "./components/UserNavBar";
 
 function App() {
@@ -322,7 +322,7 @@ function App() {
                 redirectPath="/home"
                 isAllowed={usuario.length === 1 && usuario[0].isAdmin === true}
               >
-                <PutBookID />
+                <putBookId/>
               </ProtectedRoute>
             }
           />
