+++ client/src/Styles/createAdmin.module.css
@@ -1,0 +1,21 @@
+
+.container{
+    
+    display: flex;
+  justify-content: center;
+  
+}
+.table {
+    border: 1px solid;
+  }
+
+.th {
+    border: 1px solid;
+  }
+
+.td {
+    border: 1px solid;
+  }
+
+
+
