+++ client/src/components/Admin/Admin.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react'
 import { Link } from 'react-router-dom'
-import styles from '../Styles/DashboardAdmin.module.css'
-import { getBooksAdmin , getAuthorsAdmin } from '../actions'
+import styles from '../../Styles/DashboardAdmin.module.css'
+import { getBooksAdmin , getAuthorsAdmin } from '../../actions'
 import { useDispatch } from 'react-redux'
