+++ client/src/components/Admin/AdminSearchBarBooks.jsx
@@ -3,17 +3,15 @@ import React from 'react';
 import {useState} from 'react';
 import {useDispatch} from 'react-redux';
 import {useNavigate} from 'react-router-dom';
-import {getBookTitleAdmin} from '../actions';
-import styles from '../Styles/searchBar.module.css'
+import {getBookTitleAdmin} from '../../actions';
+import styles from '../../Styles/searchBar.module.css'
 import {scroller} from "react-scroll";
-import { getUserName } from '../actions';
 
-
-const AdminSearchBarUser = () => {
+const AdminSearchBarBooks = () => {
 
     const [input, setInput] = useState('');
     const dispatch = useDispatch();
-    
+    const navigate = useNavigate();
 
     const handleChange = (e) => {
         setInput(e.target.value);
@@ -21,7 +19,7 @@ const AdminSearchBarUser = () => {
 
     const handleSubmit = (e) => {
         e.preventDefault();
-        dispatch(getUserName(input));
+        dispatch(getBookTitleAdmin(input));
         scroller.scrollTo("gaston");
         setInput('');
     }
@@ -29,11 +27,11 @@ const AdminSearchBarUser = () => {
     return (
         <div className={styles.container}>
 
-            <input type="text" placeholder='Name' value={input} onChange={(e) => handleChange(e)} className={styles.input}/>
+            <input type="text" placeholder='Título' value={input} onChange={(e) => handleChange(e)} className={styles.input}/>
             <button type='submit' onClick={(e) => handleSubmit(e)} className={styles.button}>Buscar</button>
       
         </div>
     )
 }
 
-export default AdminSearchBarUser;
+export default AdminSearchBarBooks;
