+++ client/src/components/Admin/AdminUserProfile.jsx
@@ -1,41 +1,8 @@
-// import { useDispatch } from "react-redux";
-// import { setUserPlan ,getUsers } from "../actions";
-
-
-// export default function AdminUserProfile({users , setChanged , changed}) {
-//   const dispatch = useDispatch();
-
-//   var userIds = [];
-
-//   function toogleAdmin(e, users) {
-//     users.map(usuario => {
-//       userIds.push(  usuario._id );
-//     });
-//     setTimeout(function(){
-//         dispatch(setUserPlan(userIds))
-//     }, 5000);
-    
-    
-//     dispatch(getUsers())
-//     setChanged(!changed);
-//    // console.log("changed AHORA ES es: ", changed);
-//     userIds = [];
-//   }
-
-//   return (
-//     <div>
-//       <button onClick={(e) => toogleAdmin(e, users)}>
-//        Ver Perfil
-//       </button>
-//     </div>
-//   );
-// }
-
 
 import React from "react";
 import { useSelector } from "react-redux";
 import { useParams } from "react-router-dom";
-import styles from "../Styles/adminUserProfile.module.css"
+import styles from "../../Styles/adminUserProfile.module.css"
 
 
 export default function AdminUserProfile(){
