+++ client/src/components/Admin/NavBarAdmin.jsx
@@ -1,13 +1,13 @@
 
 import React, {useState} from 'react';
-import SearchBar from './SearchBar';
-import style from '../Styles/navAdmin.module.css';
+import SearchBar from '../SearchBar';
+import style from '../../Styles/navAdmin.module.css';
 import {Link, useNavigate} from 'react-router-dom';
 import {useDispatch} from 'react-redux';
-import {getBookGenre, getBooks} from '../actions';
+import { getBookGenre , getBooks } from '../../actions';
 import {scroller} from "react-scroll";
-import LogInButton from './LogIn';
-import LogOutButton from './LogOut';
+import LogInButton from '../LogIn';
+import LogOutButton from '../LogOut';
 import { useAuth0 } from '@auth0/auth0-react';
 import { useSelector } from 'react-redux';
